Usint std::uint8_t 0.320 Mb per second.

+++ kuznechik/block.cpp
@@ -5,7 +5,7 @@ Block block_from_string(const std::string& in_string) {
     assert(in_string.length() == 128);
     Block result;
     for (std::size_t i = 0; i < 16; ++i) {
-        result[i] = static_cast<unsigned short>(std::bitset<8>(in_string.substr(8 * (15 - i), 8)).to_ulong());
+        result[i] = static_cast<std::uint8_t>(std::bitset<8>(in_string.substr(8 * (15 - i), 8)).to_ulong());
     }
     return result;
 }
@@ -14,7 +14,7 @@ Key key_from_string(const std::string& in_string) {
     assert(in_string.length() == 256);
     Key result;
     for (std::size_t i = 0; i < 32; ++i) {
-        result[i] = static_cast<unsigned short>(std::bitset<8>(in_string.substr(8 * (31 - i), 8)).to_ulong());
+        result[i] = static_cast<std::uint8_t>(std::bitset<8>(in_string.substr(8 * (31 - i), 8)).to_ulong());
     }
     return result;
 }
